+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Version 0.2 (2016-12-??)
 
-  *
+  * Implemented slice (Fixed #3)
 
 # Version 0.1 (2016-12-10)